feat: Project restructuring.
