+++ src/Chrono/Features/Audit/AuditableEntitySaveChangesInterceptor.cs
@@ -1,4 +1,6 @@
 using Chrono.Common;
+using Chrono.Entities;
+using Chrono.Entities.Common;
 using Chrono.Features.Users;
 using Chrono.Infrastructure.Persistence;
 using Microsoft.EntityFrameworkCore;
