+++ src/Chrono/Features/TaskLists/GetTaskList.cs
@@ -1,7 +1,8 @@
 using Chrono.Common.Api;
 using Chrono.Common.Exceptions;
 using Chrono.Common.Interfaces;
-using Chrono.Features.Audit;
+using Chrono.Entities;
+using Chrono.Entities.Common;
 using Chrono.Features.Tasks;
 using Chrono.Features.Users;
 using MediatR;
