+++ src/Chrono/Features/Tasks/CreateTask.cs
@@ -1,15 +1,16 @@
 using Chrono.Common.Api;
 using Chrono.Common.Exceptions;
 using Chrono.Common.Interfaces;
-using Chrono.Features.Audit;
+using Chrono.Entities;
+using Chrono.Entities.Common;
 using Chrono.Features.Categories;
-using Chrono.Features.TaskLists;
 using Chrono.Features.Users;
 using FluentValidation;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Task = Chrono.Entities.Task;
 
 namespace Chrono.Features.Tasks;
