+++ src/Chrono/Features/Tasks/TaskCategoryConfiguration.cs
@@ -1,3 +1,4 @@
+using Chrono.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
