+++ src/Chrono/Features/Users/UserConfiguration.cs
@@ -1,3 +1,4 @@
+using Chrono.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
